+++ wasatch/FeatureIdentificationDevice.py
@@ -23,6 +23,7 @@ from .StatusMessage        import StatusMessage
 from .RealUSBDevice        import RealUSBDevice
 from .MockUSBDevice        import MockUSBDevice
 from .DetectorROI          import DetectorROI
+from .PollStatus           import PollStatus
 from .EEPROM               import EEPROM
 
 log = logging.getLogger(__name__)
@@ -150,7 +151,7 @@ class FeatureIdentificationDevice(InterfaceDevice):
                 else:
                     responses.append(proc_func(*request.args, **request.kwargs))
             except Exception as e:
-                log.error(f"error in handling request {request} of {e}")
+                log.error(f"error in handling request {request} of {e}", exc_info=1)
                 responses.append(SpectrometerResponse(error_msg="error processing cmd", error_lvl=ErrorLevel.medium))
         return responses
 
@@ -245,6 +246,11 @@ class FeatureIdentificationDevice(InterfaceDevice):
         Perform additional setup after instantiating FID device.
         Split-out from physical / bus connect() to simplify MockSpectrometer.
         """
+
+        # grab firmware versions early
+        self.get_microcontroller_firmware_version()
+        self.get_fpga_firmware_version()
+        self.get_microcontroller_serial_number()
 
         # ######################################################################
         # model-specific settings
@@ -633,10 +639,13 @@ class FeatureIdentificationDevice(InterfaceDevice):
                         # for now, draw a line between previous and next_good pixels
                         # TODO: consider some kind of curve-fit
                         delta = float(spectrum[next_good] - spectrum[prev_good])
-                        rng   = next_good - prev_good
+                        rng   = next_good - prev_good 
                         step  = delta / rng
+                        # log.debug(f"correct_bad_pixels: bad_pix {bad_pix}, prev_good {prev_good} ({spectrum[prev_good]}), next_good {next_good} ({spectrum[next_good]}), delta {delta:.2f}, rng {rng}, step {step:.2f}")
                         for j in range(rng - 1):
-                            spectrum[prev_good + j + 1] = spectrum[prev_good] + step * (j + 1)
+                            new = spectrum[prev_good] + step * (j + 1)
+                            spectrum[prev_good + j + 1] = new
+                            # log.debug(f"correct_bad_pixels: stomping pix {prev_good + j + 1} with {new}")
                     else:
                         # we ran off the high end, so copy-right
                         for j in range(bad_pix, pixels):
@@ -1008,6 +1017,10 @@ class FeatureIdentificationDevice(InterfaceDevice):
     # @todo we should probably track runtime hardware gain in SpectrometerState,
     #       not EEPROM
     def set_detector_gain(self, gain: float):
+        if not self.is_sensor_stable():
+            log.error("declining to change sensor settings while stabilizing")
+            return
+
         raw = self.settings.eeprom.float_to_uint16(gain)
 
         # MZ: note that we SEND gain MSB-LSB, but we READ gain LSB-MSB?!
@@ -1056,15 +1069,22 @@ class FeatureIdentificationDevice(InterfaceDevice):
         return SpectrometerResponse(data=value)
 
     def get_microcontroller_firmware_version(self):
+        if self.settings.microcontroller_firmware_version is not None:
+            return SpectrometerResponse(data=self.settings.microcontroller_firmware_version)
+
         res = self._get_code(0xc0, label="GET_CODE_REVISION")
         result = res.data
         version = "?.?.?.?"
         if result is not None and len(result) >= 4:
             version = "%d.%d.%d.%d" % (result[3], result[2], result[1], result[0]) # MSB-LSB
+
         self.settings.microcontroller_firmware_version = version
         return SpectrometerResponse(data=version)
 
     def get_fpga_firmware_version(self):
+        if self.settings.fpga_firmware_version is not None:
+            return SpectrometerResponse(data=self.settings.fpga_firmware_version)
+
         s = ""
         res = self._get_code(0xb4, wLength=7, label="GET_FPGA_REV")
         result = res.data
@@ -1073,8 +1093,34 @@ class FeatureIdentificationDevice(InterfaceDevice):
                 c = result[i]
                 if 0x20 <= c < 0x7f: # visible ASCII
                     s += chr(c)
+
         self.settings.fpga_firmware_version = s
         return SpectrometerResponse(data=s)
+
+    def get_microcontroller_serial_number(self):
+        return None # disabling until we work out another Beta SiG conflict
+
+        if not self.settings.is_arm():
+            log.debug("GET_MICROCONTROLLER_SERIAL_NUMBER requires ARM")
+            return None
+
+        if not self.settings.supports_feature("microcontroller_serial_number"):
+            log.debug("GET_MICROCONTROLLER_SERIAL_NUMBER not supported on this firmware")
+            return None
+
+        result = self._get_code(0xff, wValue=0x2c, wLength=12, label="GET_MICROCONTROLLER_SERIAL_NUMBER")
+        if result is None:
+            log.error("GET_MICROCONTROLLER_SERIAL_NUMBER returned None")
+            return None
+
+        data = result.data
+        if data is None or len(data) != 12:
+            log.error("GET_MICROCONTROLLER_SERIAL_NUMBER expected 12 bytes (received {len(data)})")
+            return None
+
+        des = "".join(f"{v:02x}" for v in data) # Device Electronic Signature
+        self.settings.microcontroller_serial_number = des
+        return SpectrometerResponse(data=des)
 
     def apply_edc(self, spectrum):
         """
@@ -1105,6 +1151,10 @@ class FeatureIdentificationDevice(InterfaceDevice):
         if len(spectrum) != 1952:
             log.error("IMX EDC hard-coded to 1952px")
             return spectrum
+
+        # @todo: from FPGA 01.4.09+, we should be able to use spectrum [0:4] as 
+        #        TRUE optical dark; also, the "shelf" should be fixed, so the
+        #        entire left edge becomes usable.
 
         # this averages electrical dark over SPACE
         electrical_dark = sum(spectrum[1946:1950]) / 4.0 
@@ -1138,11 +1188,19 @@ class FeatureIdentificationDevice(InterfaceDevice):
                  when NOT in external-triggered mode
         @throws exception on timeout (unless external triggering enabled)
         """
+        response = SpectrometerResponse()
+
+        if not self.is_sensor_stable():
+            response.error_msg = "get_line: leaving sensor alone while stabilizing"
+            response.error_lvl = ErrorLevel.low
+            response.keep_alive = True
+            log.debug(response.error_msg)
+            self.queue_message("marquee_info", "sensor is stabilizing")
+            return response
 
         ########################################################################
         # send the ACQUIRE
         ########################################################################
-        response = SpectrometerResponse()
 
         # main use-case for NOT sending a trigger would be when reading
         # subsequent lines of data from area scan "fast" mode
@@ -1455,6 +1513,10 @@ class FeatureIdentificationDevice(InterfaceDevice):
 
         @todo SiG needs to wait 20ms + 8 frames for stabilization.
         """
+        if not self.is_sensor_stable():
+            log.error("declining to change sensor settings while stabilizing")
+            return
+
         ms = max(1, int(round(ms)))
 
         lsw =  ms        & 0xffff
@@ -1469,6 +1531,27 @@ class FeatureIdentificationDevice(InterfaceDevice):
         if self.settings.is_xs():
             self.queue_message("marquee_info", "sensor is stabilizing")
 
+        return result
+
+    def is_sensor_stable(self):
+        if not (self.settings.is_imx() and self.settings.supports_feature("imx_stabilization")):
+            return True
+
+        self.get_poll_status()
+        stable = PollStatus.IDLE == self.settings.state.poll_status
+        log.debug(f"is_sensor_stable: sensor {'IS' if stable else 'IS NOT'} stable (poll_status 0x{self.settings.state.poll_status:02x})")
+        return stable
+
+    def get_poll_status(self):
+        status = PollStatus.IDLE # .UNDEFINED (kludge)
+        result = SpectrometerResponse(data=status)
+
+        if False and self.settings.is_imx() and self.settings.supports_feature("imx_stabilization"):
+            result = self._get_code(0xd4, lsb_len=1, label="GET_POLL_STATUS")
+            if result is not None:
+                status = result.data
+        self.settings.state.poll_status = status
+        log.debug(f"get_poll_status: status 0x{status:02x}")
         return result
 
     # ##########################################################################